Revert "Steel.Effect.Atomic: disambiguate type in equality"
This is not actually needed. It did fail while I was working
on F* PR https://github.com/FStarLang/FStar/pull/3841, but after a few
changes, the original code behaves as before.

The equality is indeed between two different types, but they are
refinements of the same base type. The unifier should just join the
refinements.

This reverts commit fc35ec46065f46aa701145a2a312dc2a76002bf8.