Remove uses of term_eq_old and make sure to use new lax_term_eq
This function has been deprecated for a while and should be removed. The
old builtin term_eq had an interesting behavior of ignoring universe
annotations and ascriptions, which Steel seem to rely on. F* added a new
laxer term_eq function (https://github.com/FStarLang/FStar/pull/3936)
that mimics this old behavior, but is completely implemented in
userspace.